requirements.txt
kafka-python>=2.0.1
python-daemon>=2.2.4
